Moved game logic to board logic. Added checkForWin to board.
